+++ src/widgets/about/ui/FeatureSection.tsx
