+++ src/widgets/about/ui/HeroSection.tsx
